move tests from solana-hasher to solana-hash
