+++ sdk/hash/Cargo.toml
@@ -24,6 +24,9 @@ solana-frozen-abi = { workspace = true, optional = true }
 solana-frozen-abi-macro = { workspace = true, optional = true }
 solana-sanitize = { workspace = true }
 
+[dev-dependencies]
+solana-hash = { path = ".", features = ["dev-context-only-utils"] }
+
 [target.'cfg(target_arch = "wasm32")'.dependencies]
 js-sys = { workspace = true }
 wasm-bindgen = { workspace = true }
