+++ sdk/hasher/Cargo.toml
@@ -28,8 +28,5 @@ solana-define-syscall = { workspace = true }
 [features]
 sha2 = ["dep:sha2"]
 
-[dev-dependencies]
-bs58 = { workspace = true }
-
 [lints]
 workspace = true
